SC project/requirements.txt
streamlit
pillow
numpy
matplotlib
